Split CSS::Symbol/CSS::SymbolRaw out of CSSPrimitiveNumericTypes.h https://bugs.webkit.org/show_bug.cgi?id=283491
Reviewed by Tim Nguyen.

Split strictly non-numeric CSS::Symbol/CSS::SymbolRaw types out of the numerics file.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+SymbolDefinitions.h:
* Source/WebCore/css/values/primitives/CSSPrimitiveNumericTypes+ComputedStyleDependencies.h:
* Source/WebCore/css/values/primitives/CSSPrimitiveNumericTypes+Serialization.cpp:
* Source/WebCore/css/values/primitives/CSSPrimitiveNumericTypes+Serialization.h:
* Source/WebCore/css/values/primitives/CSSPrimitiveNumericTypes.h:
* Source/WebCore/css/values/primitives/CSSSymbol.cpp: Added.
* Source/WebCore/css/values/primitives/CSSSymbol.h: Added..

Canonical link: https://commits.webkit.org/286941@main
