+++ Source/WebCore/css/parser/CSSPropertyParserConsumer+SymbolDefinitions.h
@@ -25,10 +25,9 @@
 #pragma once
 
 #include "CSSParserToken.h"
-#include "CSSPrimitiveNumericTypes.h"
 #include "CSSPropertyParserConsumer+MetaConsumerDefinitions.h"
+#include "CSSSymbol.h"
 #include <optional>
-#include <wtf/Brigand.h>
 
 namespace WebCore {
